@@ -13,29 +13,29 @@ import { connect } from '../../utils';
 class RangeInput extends Component {
 	constructor(props) {
 		super(props);
+
+		const value = props.selectedValue || props.value || props.defaultValue || props.range;
 		this.state = {
-			start: this.props.defaultSelected
-				? this.props.defaultSelected.start
-				: props.range.start,
-			end: this.props.defaultSelected ? this.props.defaultSelected.end : props.range.end,
+			start: value.start,
+			end: value.end,
 			isStartValid: true,
 			isEndValid: true,
 		};
 	}
 
-	componentWillReceiveProps(nextProps) {
-		if (!isEqual(this.props.defaultSelected, nextProps.defaultSelected)) {
-			this.handleSlider(nextProps.defaultSelected);
+	componentDidUpdate(prevProps) {
+		if (!isEqual(this.props.value, prevProps.value)) {
+			this.handleSlider(this.props.value);
 		}
 	}
 
 	// for SSR
 	static defaultQuery = RangeSlider.defaultQuery;
-
 	static parseValue = RangeSlider.parseValue;
 
 	handleInputChange = (e) => {
 		const { name, value } = e.target;
+
 		if (Number.isNaN(value)) {
 			// set errors for invalid inputs
 			if (name === 'start') {
@@ -60,10 +60,31 @@ class RangeInput extends Component {
 				});
 			}
 		}
-		this.setState({
+
+		const currentValue = {
+			start: this.state.start,
+			end: this.state.end,
 			[name]: value,
-		});
+		};
+
+		const { value: valueProp, onChange } = this.props;
+		if (valueProp) {
+			if (onChange) onChange(currentValue);
+		} else {
+			this.setState({
+				[name]: value,
+			});
+		}
 	};
+
+	handleSliderChange = ([start, end]) => {
+		const { value, onChange } = this.props;
+		if (value) {
+			if (onChange) onChange({ start, end });
+		} else {
+			this.handleSlider({ start, end });
+		}
+	}
 
 	handleSlider = ({ start, end }) => {
 		this.setState({
@@ -86,13 +107,13 @@ class RangeInput extends Component {
 			<Container style={style} className={className}>
 				<RangeSlider
 					{...rest}
-					defaultSelected={{
+					value={{
 						start: this.state.isStartValid
 							? Number(this.state.start)
 							: this.props.range.start,
 						end: this.state.isEndValid ? Number(this.state.end) : this.props.range.end,
 					}}
-					onValueChange={this.handleSlider}
+					onChange={this.handleSliderChange}
 					className={getClassName(this.props.innerClass, 'slider-container') || null}
 				/>
 				<Flex className={getClassName(this.props.innerClass, 'input-container') || null}>
@@ -135,9 +156,10 @@ class RangeInput extends Component {
 
 RangeInput.propTypes = {
 	className: types.string,
-	defaultSelected: types.range,
+	defaultValue: types.range,
 	innerClass: types.style,
 	onValueChange: types.func,
+	onChange: types.func,
 	range: types.range,
 	stepValue: types.number,
 	style: types.style,
