@@ -196,7 +196,9 @@ class MultiDropdownList extends Component {
 					let should = [
 						{
 							[type]: {
-								[props.dataField]: value.filter(item => item !== props.missingLabel),
+								[props.dataField]: value.filter(
+									item => item !== props.missingLabel,
+								),
 							},
 						},
 					];
@@ -374,6 +376,10 @@ class MultiDropdownList extends Component {
 		const { isLastBucket } = this.state;
 		let selectAll = [];
 
+		if (this.props.isLoading && this.props.loader) {
+			return this.props.loader;
+		}
+
 		if (this.state.options.length === 0) {
 			return null;
 		}
@@ -385,7 +391,6 @@ class MultiDropdownList extends Component {
 				},
 			];
 		}
-
 		return (
 			<Container style={this.props.style} className={this.props.className}>
 				{this.props.title && (
@@ -444,6 +449,8 @@ MultiDropdownList.propTypes = {
 	value: types.stringArray,
 	filterLabel: types.string,
 	innerClass: types.style,
+	isLoading: types.bool,
+	loader: types.title,
 	onQueryChange: types.func,
 	onValueChange: types.func,
 	onChange: types.func,
@@ -495,6 +502,7 @@ const mapStateToProps = (state, props) => ({
 		(state.selectedValues[props.componentId]
 			&& state.selectedValues[props.componentId].value)
 		|| null,
+	isLoading: state.isLoading[props.componentId],
 	themePreset: state.config.themePreset,
 });
 
